requirements.txt
asgiref==3.8.1
attrs==24.3.0
certifi==2024.8.30
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.2.0
Django==5.1.1
django-cors-headers==4.4.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-yasg==1.21.7
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.1
requests==2.31.0
rpds-py==0.22.3
setuptools==75.8.0
simplejson==3.19.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
asgiref==3.8.1
attrs==24.3.0
certifi==2024.8.30
charset-normalizer==3.3.2
cloudinary==1.44.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.2.0
Django==5.1.1
django-cors-headers==4.4.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-yasg==1.21.7
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.1
requests==2.31.0
rpds-py==0.22.3
setuptools==75.8.0
simplejson==3.19.3
six==1.17.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
asgiref==3.8.1
attrs==24.3.0
cachetools==5.5.2
certifi==2024.8.30
charset-normalizer==3.3.2
cloudinary==1.44.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.2.0
Django==5.1.1
django-cors-headers==4.4.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-yasg==1.21.7
google-auth==2.40.1
google-auth-oauthlib==1.2.2
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.9.0
python-dotenv==1.0.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.1
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.22.3
rsa==4.9.1
setuptools==75.8.0
simplejson==3.19.3
six==1.17.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
asgiref==3.8.1
attrs==24.3.0
cachetools==5.5.2
certifi==2024.8.30
charset-normalizer==3.3.2
cloudinary==1.44.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.2.0
Django==5.1.1
django-allauth==65.8.0
django-cors-headers==4.4.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-yasg==1.21.7
google-auth==2.40.1
google-auth-oauthlib==1.2.2
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.9.0
python-dotenv==1.0.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.1
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.22.3
rsa==4.9.1
setuptools==75.8.0
simplejson==3.19.3
six==1.17.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
asgiref==3.8.1
attrs==24.3.0
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cloudinary==1.44.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.3
dj-database-url==2.2.0
Django==5.1.1
django-allauth==65.8.0
django-cors-headers==4.4.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-yasg==1.21.7
google-auth==2.40.1
google-auth-oauthlib==1.2.2
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.1
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.22.3
rsa==4.9.1
setuptools==75.8.0
simplejson==3.19.3
six==1.17.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
